+++ pkg/network/ebpf/c/co-re/tracer-fentry.c
@@ -8,6 +8,7 @@
 #include "ipv6.h"
 #include "sock.h"
 #include "skb.h"
+#include "task_event.h"
 
 #include "tracer/tracer.h"
 #include "tracer/events.h"
@@ -20,37 +21,6 @@
 
 BPF_PERCPU_HASH_MAP(udp6_send_skb_args, u64, u64, 1024)
 BPF_PERCPU_HASH_MAP(udp_send_skb_args, u64, conn_tuple_t, 1024)
-
-#define RETURN_IF_NOT_IN_SYSPROBE_TASK(prog_name)           \
-    if (!event_in_task(prog_name)) {                        \
-        return 0;                                           \
-    }
-
-static __always_inline __u32 systemprobe_dev() {
-    __u64 val = 0;
-    LOAD_CONSTANT("systemprobe_device", val);
-    return (__u32)val;
-}
-
-static __always_inline __u32 systemprobe_ino() {
-    __u64 val = 0;
-    LOAD_CONSTANT("systemprobe_ino", val);
-    return (__u32)val;
-}
-
-static __always_inline bool event_in_task(char *prog_name) {
-    __u32 dev = systemprobe_dev();
-    __u32 ino = systemprobe_ino();
-    struct bpf_pidns_info ns = {};
-
-    u64 error = bpf_get_ns_current_pid_tgid(dev, ino, &ns, sizeof(struct bpf_pidns_info));
-
-    if (error) {
-        log_debug("%s: err=event originates from outside current fargate task\n", prog_name);
-    }
-
-    return !error;
-}
 
 static __always_inline int read_conn_tuple_partial_from_flowi4(conn_tuple_t *t, struct flowi4 *fl4, u64 pid_tgid, metadata_mask_t type) {
     t->pid = pid_tgid >> 32;
